requirements.txt
alembic==1.8.1
astroid==2.11.6
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
dill==0.3.5.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
graphql-core==3.2.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.1
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pydantic==1.9.1
pylint==2.14.4
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.39
tomli==2.0.1
tomlkit==0.11.0
typing-extensions==4.2.0
urllib3==1.26.9
validators==0.20.0
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
